+++ pincer/utils/__init__.py
@@ -6,7 +6,7 @@ from .conversion import convert
 from .extraction import get_index, get_signature_and_params
 from .insertion import should_pass_cls, should_pass_ctx
 from .snowflake import Snowflake
-from .task import Task, TaskScheduler
+from .tasks import Task, TaskScheduler
 from .timestamp import Timestamp
 from .types import APINullable, Coro, MISSING, Choices, choice_value_types
